+++ examples/api/src/main.js
@@ -2,12 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-License-Identifier: MIT
 
-import "uno.css";
-import "./app.css";
-import App from "./App.svelte";
+import 'uno.css'
+import './app.css'
+import App from './App.svelte'
 
 const app = new App({
-  target: document.querySelector("#app"),
-});
+  target: document.querySelector('#app')
+})
 
-export default app;
+export default app