+++ plugins/deep-link/examples/app/src-tauri/src/lib.rs
@@ -14,6 +14,11 @@ fn greet(name: &str) -> String {
 pub fn run() {
     tauri::Builder::default()
         .plugin(tauri_plugin_deep_link::init())
+        .plugin(
+            tauri_plugin_log::Builder::default()
+                .level(log::LevelFilter::Info)
+                .build(),
+        )
         .setup(|app| {
             app.listen("deep-link://new-url", |url| {
                 dbg!(url);