+++ plugins/dialog/src/commands.rs
@@ -136,7 +136,7 @@ pub(crate) async fn open<R: Runtime>(
                 let folders = dialog_builder.blocking_pick_folders();
                 if let Some(folders) = &folders {
                     for folder in folders {
-                        if let Ok(path) = folder.path() {
+                        if let Ok(path) = folder.clone().into_path() {
                             if let Some(s) = window.try_fs_scope() {
                                 s.allow_directory(path, options.recursive);
                             }
@@ -149,7 +149,7 @@ pub(crate) async fn open<R: Runtime>(
             } else {
                 let folder = dialog_builder.blocking_pick_folder();
                 if let Some(folder) = &folder {
-                    if let Ok(path) = folder.path() {
+                    if let Ok(path) = folder.clone().into_path() {
                         if let Some(s) = window.try_fs_scope() {
                             s.allow_directory(path, options.recursive);
                         }
@@ -164,7 +164,7 @@ pub(crate) async fn open<R: Runtime>(
         let files = dialog_builder.blocking_pick_files();
         if let Some(files) = &files {
             for file in files {
-                if let Ok(path) = file.path() {
+                if let Ok(path) = file.clone().into_path() {
                     if let Some(s) = window.try_fs_scope() {
                         s.allow_file(&path);
                     }
@@ -178,7 +178,7 @@ pub(crate) async fn open<R: Runtime>(
         let file = dialog_builder.blocking_pick_file();
 
         if let Some(file) = &file {
-            if let Ok(path) = file.path() {
+            if let Ok(path) = file.clone().into_path() {
                 if let Some(s) = window.try_fs_scope() {
                     s.allow_file(&path);
                 }
@@ -197,41 +197,36 @@ pub(crate) async fn save<R: Runtime>(
     dialog: State<'_, Dialog<R>>,
     options: SaveDialogOptions,
 ) -> Result<Option<FilePath>> {
-    #[cfg(target_os = "ios")]
-    return Err(crate::Error::FileSaveDialogNotImplemented);
-    #[cfg(any(desktop, target_os = "android"))]
+    let mut dialog_builder = dialog.file();
+    #[cfg(any(windows, target_os = "macos"))]
     {
-        let mut dialog_builder = dialog.file();
-        #[cfg(any(windows, target_os = "macos"))]
-        {
-            dialog_builder = dialog_builder.set_parent(&window);
-        }
-        if let Some(title) = options.title {
-            dialog_builder = dialog_builder.set_title(title);
-        }
-        if let Some(default_path) = options.default_path {
-            dialog_builder = set_default_path(dialog_builder, default_path);
-        }
-        if let Some(can) = options.can_create_directories {
-            dialog_builder = dialog_builder.set_can_create_directories(can);
-        }
-        for filter in options.filters {
-            let extensions: Vec<&str> = filter.extensions.iter().map(|s| &**s).collect();
-            dialog_builder = dialog_builder.add_filter(filter.name, &extensions);
-        }
-
-        let path = dialog_builder.blocking_save_file();
-        if let Some(p) = &path {
-            if let Ok(path) = p.path() {
-                if let Some(s) = window.try_fs_scope() {
-                    s.allow_file(&path);
-                }
-                window.state::<tauri::scope::Scopes>().allow_file(&path)?;
-            }
-        }
-
-        Ok(path.map(|p| p.simplified()))
+        dialog_builder = dialog_builder.set_parent(&window);
     }
+    if let Some(title) = options.title {
+        dialog_builder = dialog_builder.set_title(title);
+    }
+    if let Some(default_path) = options.default_path {
+        dialog_builder = set_default_path(dialog_builder, default_path);
+    }
+    if let Some(can) = options.can_create_directories {
+        dialog_builder = dialog_builder.set_can_create_directories(can);
+    }
+    for filter in options.filters {
+        let extensions: Vec<&str> = filter.extensions.iter().map(|s| &**s).collect();
+        dialog_builder = dialog_builder.add_filter(filter.name, &extensions);
+    }
+
+    let path = dialog_builder.blocking_save_file();
+    if let Some(p) = &path {
+        if let Ok(path) = p.clone().into_path() {
+            if let Some(s) = window.try_fs_scope() {
+                s.allow_file(&path);
+            }
+            window.state::<tauri::scope::Scopes>().allow_file(&path)?;
+        }
+    }
+
+    Ok(path.map(|p| p.simplified()))
 }
 
 fn message_dialog<R: Runtime>(